completed quering and storing stock price

+++ database/stockSchema.sql
@@ -8,7 +8,7 @@ CREATE TABLE stock (
   id INT NOT NULL AUTO_INCREMENT,
   stock_ticker VARCHAR(20) NOT NULL,
   quantity INT NOT NULL,
-  price INT NOT NULL,
+  price FLOAT(7, 2) NOT NULL,
   UNIQUE (stock_ticker),
   PRIMARY KEY (id)
 );
